% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{demo_c}
\alias{demo_c}
\title{Demographics Data from NHANES 2003-2004}
\source{
\url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}
}
\description{
Includes 44 variables.
}
