% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{demo_d}
\alias{demo_d}
\title{Demographics Data from NHANES 2005-2006}
\source{
\url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2005}
}
\description{
Includes 43 variables.
}
