% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{ocq_c}
\alias{ocq_c}
\title{Occupational Questionnaire Data from NHANES 2003-2004}
\source{
\url{https://wwwn.cdc.gov/Nchs/Nhanes/2003-2004/OCQ_C.htm}
}
\description{
Includes 25 variables.
}
