% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{ocq_d}
\alias{ocq_d}
\title{Occupational Questionnaire Data from NHANES 2005-2006}
\source{
\url{https://wwwn.cdc.gov/Nchs/Nhanes/2005-2006/OCQ_D.htm}
}
\description{
Includes 15 variables.
}
