% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nhanes_app.R
\name{process_nhanes_app}
\alias{process_nhanes_app}
\title{Shiny App to Process NHANES 2003-2006 Accelerometer Data}
\usage{
process_nhanes_app()
}
\description{
Simply a Shiny app that provides a GUI for \code{\link{process_nhanes}}.
}
