% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_nhanes.R
\name{reweight_nhanes}
\alias{reweight_nhanes}
\title{Calculate Adjusted Mobile Exam Center (MEC) Weights}
\usage{
reweight_nhanes(accel_data, wave, seqn_column, include_column)
}
\arguments{
\item{accel_data}{Numeric matrix or data frame with one row per participant.
It can have any number of columns but one must have respondent sequence
numbers (NHANES ID variable "seqn") and another must indicate whether each
participant has valid data (1 if valid, 0 if not valid).}

\item{wave}{Integer value specifying NHANES wave. Choices are 1 for NHANES
2003-2004 and 2 for NHANES 2005-2006.}

\item{seqn_column}{Integer value specifying column number for respondent
sequence numbers (NHANES ID variable "seqn").}

\item{include_column}{Integer value specifying column number for valid day
variable, which has 1's for valid days and 0's for invalid days.}
}
\value{
Numeric matrix identical to \code{accel_data}, but with extra
\code{wtmec2yr_adj} column.
}
\description{
Calculates adjusted Mobile Exam Center (MEC) weights for the subsample of
NHANES 2003-2004 and 2005-2006 participants with valid accelerometer data,
making the subsample nationally representative. This function is used
internally by \code{\link{process_nhanes}}, but could also be used
independently.
}
\examples{
# Process NHANES 2003-2004 data using default settings
nhanes0304 <- process_nhanes(waves = 1)

# Create new 'include' variable: 1 if >= 4 valid days, otherwise 0
nhanes0304$include2 <- ifelse(nhanes0304$valid_days >= 4, 1, 0)

# Rename previous 2-year weight variable so new one doesn't overwrite it
names(nhanes0304)[10] = "wtmec2yr_adj_old"

# Calculate new 2-year weights for subset with >= 4 valid days
nhanes0304 <- reweight_nhanes(accel_data = nhanes0304, wave = 1,
                              seqn_column = 1, include_column = 12)

# Notice that 'wtmec2yr_adj' differ from 'wtmec2yr_adj_old', and that
# participants with 1-3 valid days are now assigned zero weight
nhanes0304[1: 20, ]

# This is just an example. In practice, it would be easier to just add
# `valid.days = 4' to the initial process.nhanes function call above.


}
