\name{w1}
\alias{w1}
\docType{data}
\title{
NHANES 2003-2004 miscellaneous accelerometer data
}
\description{
Not intended for direct use.
}
\usage{data(w1)}
\format{
  The format is:
 num [1:7176, 1:7] 21005 21006 21007 21008 21009 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "seqn" "start" "stop" "stat" ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2003-2004 [1].
}
\references{
1. Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2003-4 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}. Accessed February 1, 2014.
}
\examples{
data(w1)
## maybe str(w1) ; plot(w1) ...
}
\keyword{datasets}