\name{w2}
\alias{w2}
\docType{data}
\title{
NHANES 2003-2004 miscellaneous accelerometer data
}
\description{
Not intended for direct use.
}
\usage{data(w2)}
\format{
  The format is:
 num [1:7455, 1:7] 31128 31129 31131 31132 31133 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "seqn" "start" "stop" "stat" ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2005-2006 [1].
}
\references{
1. Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2005-6 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2005}. Accessed February 1, 2014.
}
\examples{
data(w2)
## maybe str(w2) ; plot(w2) ...
}
\keyword{datasets}