\name{wave1_ages}
\alias{wave1_ages}
\docType{data}
\title{
NHANES 2003-2004 participant ages
}
\description{
Not intended for direct use.
}
\usage{data(wave1_ages)}
\format{
  The format is:
 num [1:7176] 19 16 14 17 55 52 63 13 83 13 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2003-2004 [1].
}
\references{
1. Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2003-4 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}. Accessed February 1, 2014.
}
\examples{
NA
}
\keyword{datasets}