\name{wave1_demo}
\alias{wave1_demo}
\docType{data}
\title{
NHANES 2003-2004 select variables from 'Demographics' file
}
\description{
Not intended for direct use.
}
\usage{data(wave1_demo)}
\format{
  The format is:
 num [1:10122, 1:5] 21005 21006 21007 21008 21009 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2003-2004 (see reference).
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2003-4 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}. Accessed August 16, 2013.
}
\examples{
NA
}
\keyword{datasets}