\name{wave1_paxday}
\alias{wave1_paxday}
\docType{data}
\title{
NHANES 2003-2004 accelerometer 'day' indicator
}
\description{
Not intended for direct use.
}
\usage{data(wave1_paxday)}
\format{
  The format is:
 int [1:72250027] 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2003-2004 [1].
}
\references{
1. Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2003-4 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2003}. Accessed February 1, 2014.
}
\examples{
NA
}
\keyword{datasets}