\name{wave1_seqn}
\alias{wave1_seqn}
\docType{data}
\title{
NHANES 2003-2004 respondent sequence numbers
}
\description{
Not intended for direct use.
}
\usage{data(wave1_seqn)}
\format{
  The format is:
 int [1:72250027] 21005 21005 21005 21005 21005 21005 21005 21005 21005 21005 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2003-2004 (see reference).
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2003-4 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}. Accessed August 16, 2013.
}
\examples{
NA
}
\keyword{datasets}