\name{wave2_demo}
\alias{wave2_demo}
\docType{data}
\title{
NHANES 2005-2006 select variables from 'Demographics' file
}
\description{
Not intended for direct use.
}
\usage{data(wave2_demo)}
\format{
  The format is:
 num [1:10348, 1:5] 31127 31128 31129 31130 31131 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2005-2006 (see reference).
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2005-6 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2005}. Accessed August 16, 2013.
}
\examples{
NA
}
\keyword{datasets}