\name{wave2_paxstep}
\alias{wave2_paxstep}
\docType{data}
\title{
NHANES 2005-2006 accelerometer 'step' values
}
\description{
Not intended for direct use.
}
\usage{data(wave2_paxstep)}
\format{
  The format is:
 int [1:74874095] 4 0 0 4 0 0 0 0 0 0 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2005-2006 (see reference).
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2005-6 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2005}. Accessed August 16, 2013.
}
\examples{
NA
}
\keyword{datasets}