\name{wave2_seqn}
\alias{wave2_seqn}
\docType{data}
\title{
NHANES 2005-2006 respondent sequence numbers
}
\description{
Not intended for direct use.
}
\usage{data(wave2_seqn)}
\format{
  The format is:
 int [1:74874095] 31128 31128 31128 31128 31128 31128 31128 31128 31128 31128 ...
}
\details{
NA
}
\source{
National Health and Nutrition Examination Survey 2005-2006 [1].
}
\references{
1. Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: US Department of Health and Human Services, Centers for Disease Control and Prevention, 2005-6 \url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2005}. Accessed February 1, 2014.
}
\examples{
NA
}
\keyword{datasets}