% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{dem}
\alias{dem}
\title{NHANES 2003-2006 Demographics Data}
\source{
\url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2003}
}
\description{
Includes 45 demographics variables for participants in the National Health 
and Nutrition Examination Survey 2003-2006 dataset.
}
