% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{w1}
\alias{w1}
\title{Index Data for NHANES 2003-2004 Accelerometer Data}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2003}
}
\description{
Includes start/end positions and status indicators for accelerometer data.
}
