% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{wave1_paxinten}
\alias{wave1_paxinten}
\title{Accelerometer Count Data from NHANES 2003-2004}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2003}
}
\description{
Long vector used by \code{\link{process_nhanes}}.
}
