% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{wave2_demo}
\alias{wave2_demo}
\title{Select NHANES 2005-2006 Demographics Variables}
\source{
\url{http://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Demographics&CycleBeginYear=2005}
}
\description{
Includes 5 variables used by \code{\link{process_nhanes}}.
}
