% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesaccel-data.R
\docType{data}
\name{wave2_paxstep}
\alias{wave2_paxstep}
\title{Accelerometer Steps Data from NHANES 2005-2006}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2005}
}
\description{
Long vector used by \code{\link{process_nhanes}}.
}
